Fix linting errors on moved files.
Note: Changed one instance of vjs() to videojs().  Technically both
are exposed as globals by the video.js library, but actually one
is just an alias for the other, and I'd like to avoid using globals.

Reference for videojs being an alias for vjs is the function comment
here (note this is back on version 4.12 of videojs, which we use)
https://github.com/videojs/video.js/blob/0fe811df2c357712cc55562d80588c52bb003ea8/src/js/core.js#L26